refactor: move rest of inline routes into seperate fiels
